// src/api/generated/models/searchResult.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * CRM API GetMoto
 * API para gestão de oficina de motos - Sistema de CRM completo
 * OpenAPI spec version: 1.0.0
 */
import type { SearchResultType } from './searchResultType';
import type { SearchResultMetadata } from './searchResultMetadata';

export interface SearchResult {
  id: string;
  type: SearchResultType;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  metadata?: SearchResultMetadata;
}