// src/api/generated/models/searchResultType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * CRM API GetMoto
 * API para gestão de oficina de motos - Sistema de CRM completo
 * OpenAPI spec version: 1.0.0
 */

export type SearchResultType = typeof SearchResultType[keyof typeof SearchResultType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SearchResultType = {
  product: 'product',
  service: 'service',
  vehicle: 'vehicle',
  serviceOrder: 'serviceOrder',
} as const;